+++ internal/concurrency/concurrency.go
@@ -15,3 +15,12 @@ func NewPool(ctx context.Context, maxGoroutines int) *pool.ContextPool {
 		WithFirstError().
 		WithMaxGoroutines(maxGoroutines)
 }
+
+// TrySendThroughChannel attempts to send an object through a channel.
+// If the context is canceled, it will not send the object.
+func TrySendThroughChannel[T any](ctx context.Context, msg T, channel chan<- T) {
+	select {
+	case <-ctx.Done():
+	case channel <- msg:
+	}
+}
