+++ pkg/server/commands/listusers/list_users_rpc.go
@@ -353,7 +353,7 @@ func (l *listUsersQuery) expand(
 
 	for _, userFilter := range req.GetUserFilters() {
 		if reqObjectType == userFilter.GetType() && reqRelation == userFilter.GetRelation() {
-			trySendResult(ctx, foundUser{
+			concurrency.TrySendThroughChannel(ctx, foundUser{
 				user: &openfgav1.User{
 					User: &openfgav1.User_Userset{
 						Userset: &openfgav1.UsersetUser{
@@ -496,7 +496,7 @@ LoopOnIterator:
 				if f.GetType() == userObjectType {
 					user := tuple.StringToUserProto(tuple.BuildObject(userObjectType, userObjectID))
 
-					trySendResult(ctx, foundUser{
+					concurrency.TrySendThroughChannel(ctx, foundUser{
 						user: user,
 					}, foundUsersChan)
 				}
@@ -621,7 +621,7 @@ func (l *listUsersQuery) expandIntersection(
 				user:          tuple.StringToUserProto(key),
 				excludedUsers: excludedUsers,
 			}
-			trySendResult(ctx, fu, foundUsersChan)
+			concurrency.TrySendThroughChannel(ctx, fu, foundUsersChan)
 		}
 	}
 
@@ -705,7 +705,7 @@ func (l *listUsersQuery) expandUnion(
 			user:          tuple.StringToUserProto(key),
 			excludedUsers: excludedUsers,
 		}
-		trySendResult(ctx, fu, foundUsersChan)
+		concurrency.TrySendThroughChannel(ctx, fu, foundUsersChan)
 	}
 
 	return expandResponse{
@@ -776,7 +776,7 @@ func (l *listUsersQuery) expandExclusion(
 		switch {
 		case baseWildcardExists:
 			if !userIsSubtracted && !wildcardSubtracted {
-				trySendResult(ctx, foundUser{
+				concurrency.TrySendThroughChannel(ctx, foundUser{
 					user: tuple.StringToUserProto(userKey),
 				}, foundUsersChan)
 			}
@@ -784,7 +784,7 @@ func (l *listUsersQuery) expandExclusion(
 			for subtractedUserKey, subtractedFu := range subtractFoundUsersMap {
 				if tuple.IsTypedWildcard(subtractedUserKey) {
 					if !userIsSubtracted {
-						trySendResult(ctx, foundUser{
+						concurrency.TrySendThroughChannel(ctx, foundUser{
 							user:               tuple.StringToUserProto(userKey),
 							relationshipStatus: NoRelationship,
 						}, foundUsersChan)
@@ -793,7 +793,7 @@ func (l *listUsersQuery) expandExclusion(
 				}
 
 				if subtractedFu.relationshipStatus == NoRelationship {
-					trySendResult(ctx, foundUser{
+					concurrency.TrySendThroughChannel(ctx, foundUser{
 						user:               tuple.StringToUserProto(subtractedUserKey),
 						relationshipStatus: HasRelationship,
 					}, foundUsersChan)
@@ -802,7 +802,7 @@ func (l *listUsersQuery) expandExclusion(
 				// a found user under the subtracted branch causes the subtracted user to have a negated relationship with respect
 				// to the base relation and is excluded since a wildcard is contained under the base branch.
 				if subtractedFu.relationshipStatus == HasRelationship {
-					trySendResult(ctx, foundUser{
+					concurrency.TrySendThroughChannel(ctx, foundUser{
 						user:               tuple.StringToUserProto(subtractedUserKey),
 						relationshipStatus: NoRelationship,
 						excludedUsers: []*openfgav1.User{
@@ -813,21 +813,21 @@ func (l *listUsersQuery) expandExclusion(
 			}
 		case subtractWildcardExists, userIsSubtracted:
 			if subtractedUser.relationshipStatus == HasRelationship {
-				trySendResult(ctx, foundUser{
+				concurrency.TrySendThroughChannel(ctx, foundUser{
 					user:               tuple.StringToUserProto(userKey),
 					relationshipStatus: NoRelationship,
 				}, foundUsersChan)
 			}
 
 			if subtractedUser.relationshipStatus == NoRelationship {
-				trySendResult(ctx, foundUser{
+				concurrency.TrySendThroughChannel(ctx, foundUser{
 					user:               tuple.StringToUserProto(userKey),
 					relationshipStatus: HasRelationship,
 				}, foundUsersChan)
 			}
 
 		default:
-			trySendResult(ctx, foundUser{
+			concurrency.TrySendThroughChannel(ctx, foundUser{
 				user:               tuple.StringToUserProto(userKey),
 				relationshipStatus: fu.relationshipStatus,
 			}, foundUsersChan)
@@ -946,15 +946,6 @@ func (l *listUsersQuery) buildResultsChannel() chan foundUser {
 	}
 
 	return foundUsersCh
-}
-
-func trySendResult(ctx context.Context, user foundUser, foundUsersCh chan<- foundUser) {
-	select {
-	case <-ctx.Done():
-		return
-	case foundUsersCh <- user:
-		return
-	}
 }
 
 func tupleConditionMet(ctx context.Context, reqCtx *structpb.Struct, typesys *typesystem.TypeSystem, t *openfgav1.TupleKey) (bool, error) {
